# webcalendar-openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WebCalendar API
  description: REST API for WebCalendar Next Generation (WCTNG)
  version: 2.0.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:47180/api/v2
    description: Local development
  - url: https://api.example.com/api/v2
    description: Production (placeholder)
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: List all users (admin only)
      description: Returns all users. Requires admin privileges.
      tags:
        - Users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
                  error:
                    type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      operationId: createUser
      summary: Create a new user (admin only)
      description: Creates a new user account. Requires admin privileges.
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
                  meta:
                    type: "null"
                  error:
                    type: "null"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Login already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: "null"
                  meta:
                    type: "null"
                  error:
                    $ref: "#/components/schemas/ErrorResponse"
  /users/{login}:
    get:
      operationId: getUser
      summary: Get user profile
      description: Returns a user profile. Admin can view any user; non-admin can only view self.
      tags:
        - Users
      parameters:
        - name: login
          in: path
          required: true
          schema:
            type: string
          description: User login identifier
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
                  meta:
                    type: "null"
                  error:
                    type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateUser
      summary: Update user profile
      description: Updates a user profile. Admin can update any user; non-admin can update self only.
      tags:
        - Users
      parameters:
        - name: login
          in: path
          required: true
          schema:
            type: string
          description: User login identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
                  meta:
                    type: "null"
                  error:
                    type: "null"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{login}/password:
    put:
      operationId: changePassword
      summary: Change user password
      description: Changes a user's password. Non-admin must provide current_password; admin can skip it.
      tags:
        - Users
      parameters:
        - name: login
          in: path
          required: true
          schema:
            type: string
          description: User login identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                  meta:
                    type: "null"
                  error:
                    type: "null"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    ErrorResponse:
      type: object
      description: Standard error object returned in the envelope when a request fails.
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 404
        message:
          type: string
          description: Human-readable error message
          example: Event not found
        details:
          type: array
          items:
            type: string
          description: Optional list of detailed error messages (e.g. validation errors)
          default: []
    PaginationMeta:
      type: object
      description: Pagination metadata included in list responses.
      required:
        - total
        - page
        - limit
      properties:
        total:
          type: integer
          description: Total number of items matching the query
          example: 100
        page:
          type: integer
          description: Current page number (1-based)
          example: 1
        limit:
          type: integer
          description: Maximum items per page
          example: 20
...
    User:
      type: object
      description: User account profile.
      required:
        - login
        - firstname
        - lastname
        - email
        - is_admin
        - enabled
      properties:
        login:
          type: string
          description: Unique login identifier
          example: john_doe
        firstname:
          type: string
          description: First name
          example: John
        lastname:
          type: string
          description: Last name
          example: Doe
        email:
          type: string
          format: email
          description: Email address
          example: john@example.com
        is_admin:
          type: boolean
          description: Admin privileges
          example: false
        enabled:
          type: boolean
          description: Account enabled
          example: true
    UserCreateRequest:
      type: object
      description: Request body for creating a user (admin only).
      required:
        - login
        - password
        - email
      properties:
        login:
          type: string
          example: newuser
        password:
          type: string
          format: password
        email:
          type: string
          format: email
        firstname:
          type: string
        lastname:
          type: string
        is_admin:
          type: boolean
          default: false
    UserUpdateRequest:
      type: object
      description: Request body for updating a user profile. All fields optional.
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
          format: email
        is_admin:
          type: boolean
          description: Admin only
        enabled:
          type: boolean
          description: Admin only
    PasswordChangeRequest:
      type: object
      description: Request body for changing a password.
      required:
        - new_password
      properties:
        current_password:
          type: string
          format: password
          description: Required for non-admin users
        new_password:
          type: string
          format: password
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: "null"
              meta:
                type: "null"
              error:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            meta: null
            error:
              code: 401
              message: Invalid or expired token
              details: []
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: "null"
              meta:
                type: "null"
              error:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            meta: null
            error:
              code: 404
              message: Resource not found
              details: []
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: "null"
              meta:
                type: "null"
              error:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            meta: null
            error:
              code: 403
              message: Forbidden
              details: []
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: "null"
              meta:
                type: "null"
              error:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            meta: null
            error:
              code: 400
              message: Validation failed
              details:
                - "title is required"